# Switch to using MethodHandles for property value propagations