… src/main/java/org/eclipse/yasson/internal/model/GetValueCommand.java
… src/main/java/org/eclipse/yasson/internal/model/MethodHandleValuePropagation.java
… src/main/java/org/eclipse/yasson/internal/model/PropertyValuePropagation.java
… src/main/java/org/eclipse/yasson/internal/model/ReflectionPropagation.java
… src/main/java/org/eclipse/yasson/internal/model/SetValueCommand.java
… src/test/java/org/eclipse/yasson/defaultmapping/modifiers/ClassModifiersTest.java