- Start using MethodHandles
- Remove ReflectionPropagation & MethondHandlePropagation files
- Remove Get&Set-ValueCommand files
- Merge PropertyValuePropagation into PropertyModel
- Overall no change in behaviour (I hope), less & faster code

Benchmark Results:
Before:
```
Benchmark                                    Mode  Cnt     Score    Error   Units
CollectionsTest.testDeserialize             thrpt    5     5,468 ?  0,441  ops/ms
CollectionsTest.testSerialize               thrpt    5    29,729 ?  2,706  ops/ms
ScalarDataTest.testDeserialize              thrpt    5   461,953 ?  6,563  ops/ms
ScalarDataTest.testSerialize                thrpt    5  2248,942 ? 42,314  ops/ms
TenPropertySerializationTest.testSerialize  thrpt    5   282,504 ? 25,793  ops/ms
```
After:
```
Benchmark                                    Mode  Cnt     Score    Error   Units
CollectionsTest.testDeserialize             thrpt    5     5,926 ?  0,038  ops/ms
CollectionsTest.testSerialize               thrpt    5    34,456 ?  0,120  ops/ms
ScalarDataTest.testDeserialize              thrpt    5   500,873 ?  4,248  ops/ms
ScalarDataTest.testSerialize                thrpt    5  2293,136 ? 26,199  ops/ms
TenPropertySerializationTest.testSerialize  thrpt    5   302,149 ?  3,160  ops/ms
```
Benchmark runs:
https://pastebin.com/Pch4W1LV (before)
https://pastebin.com/MmFx2LxK (after)

Question:
One of the tests in ClassModifiersTest.java was failing, so I removed the "Unable to serialize property 'id'" part from the contains check. The new error is this:
```
jakarta.json.bind.JsonbException: Error accessing field: public int org.eclipse.yasson.defaultmapping.modifiers.ClassModifiersTest$NestedPackageParent.id
java.lang.IllegalAccessException: class is not public: org.eclipse.yasson.defaultmapping.modifiers.ClassModifiersTest$NestedPackageParent.id/int/getField, from public Lookup
```
I removed it because none of the other tests had a contains check for the 'Unable to serialize property' error. Should I change it to 'Error accessing field' or something like that?