@Degubi 2020-07-06
I'm getting an error with the build:
```
The command "~/bin/install-jdk.sh --target "/home/travis/oraclejdk11" --workspace "/home/travis/.cache/install-jdk" --feature "11" --license "BCL"" failed and exited with 56 during .
```