@aguibert 2020-07-07
@Degubi yea I was having issues with CI too. I believe this is an issue on the TravisCI end and hopefully it is better now