@Degubi 2020-07-07
@aguibert Seems like it's fixed now